requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
ez-setup==0.9
idna==3.4
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
